Workaround for tooltips appearing at coord 0
Use BaseTooltip

Shorten build names
