+++ plugins/linear-genome-view/src/BaseLinearDisplay/components/Tooltip.tsx
@@ -1,37 +1,10 @@
 import React from 'react'
 import { getConf } from '@jbrowse/core/configuration'
 import { SanitizedHTML } from '@jbrowse/core/ui'
+import BaseTooltip from '@jbrowse/core/ui/BaseTooltip'
 import { observer } from 'mobx-react'
-import { Portal, alpha } from '@mui/material'
-import { makeStyles } from 'tss-react/mui'
-import {
-  useClientPoint,
-  useFloating,
-  useInteractions,
-} from '@floating-ui/react'
-
 // locals
 import { BaseLinearDisplayModel } from '../models/BaseLinearDisplayModel'
-
-function round(value: number) {
-  return Math.round(value * 1e5) / 1e5
-}
-const useStyles = makeStyles()(theme => ({
-  // these styles come from
-  // https://github.com/mui-org/material-ui/blob/master/packages/material-ui/src/Tooltip/Tooltip.js
-  tooltip: {
-    pointerEvents: 'none',
-    backgroundColor: alpha(theme.palette.grey[700], 0.9),
-    borderRadius: theme.shape.borderRadius,
-    color: theme.palette.common.white,
-    fontFamily: theme.typography.fontFamily,
-    padding: '4px 8px',
-    fontSize: theme.typography.pxToRem(12),
-    lineHeight: `${round(14 / 10)}em`,
-    maxWidth: 300,
-    wordWrap: 'break-word',
-  },
-}))
 
 interface Props {
   message: React.ReactNode | string
@@ -58,40 +31,18 @@ const Tooltip = observer(function ({
   model: BaseLinearDisplayModel
   clientMouseCoord: Coord
 }) {
-  const { theme, classes } = useStyles()
   const { featureUnderMouse } = model
-
   const x = clientMouseCoord[0] + 15
   const y = clientMouseCoord[1]
-
-  const { refs, floatingStyles, context } = useFloating({
-    placement: 'right',
-    strategy: 'fixed',
-  })
-
-  const clientPoint = useClientPoint(context, { x, y })
-  const { getFloatingProps } = useInteractions([clientPoint])
 
   const contents = featureUnderMouse
     ? getConf(model, 'mouseover', { feature: featureUnderMouse })
     : undefined
 
-  const popperTheme = theme.components?.MuiPopper
   return featureUnderMouse && contents ? (
-    <Portal container={popperTheme?.defaultProps?.container}>
-      <div
-        className={classes.tooltip}
-        ref={refs.setFloating}
-        style={{
-          ...floatingStyles,
-          zIndex: 100000,
-          pointerEvents: 'none',
-        }}
-        {...getFloatingProps()}
-      >
-        <TooltipContents message={contents} />
-      </div>
-    </Portal>
+    <BaseTooltip clientPoint={{ x, y }}>
+      <TooltipContents message={contents} />
+    </BaseTooltip>
   ) : null
 })
