+++ plugins/wiggle/src/Tooltip.tsx
@@ -1,33 +1,12 @@
 import React from 'react'
 import { observer } from 'mobx-react'
-import { alpha, Portal, useTheme } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
 import { Feature } from '@jbrowse/core/util'
-import {
-  useClientPoint,
-  useFloating,
-  useInteractions,
-} from '@floating-ui/react'
+import BaseTooltip from '@jbrowse/core/ui/BaseTooltip'
 // locals
-import { YSCALEBAR_LABEL_OFFSET, round } from './util'
+import { YSCALEBAR_LABEL_OFFSET } from './util'
 
-const useStyles = makeStyles()(theme => ({
-  // these styles come from
-  // https://github.com/mui-org/material-ui/blob/master/packages/material-ui/src/Tooltip/Tooltip.js
-  tooltip: {
-    position: 'absolute',
-    pointerEvents: 'none',
-    backgroundColor: alpha(theme.palette.grey[700], 0.9),
-    borderRadius: theme.shape.borderRadius,
-    color: theme.palette.common.white,
-    fontFamily: theme.typography.fontFamily,
-    padding: '4px 8px',
-    fontSize: theme.typography.pxToRem(12),
-    lineHeight: `${round(14 / 10)}em`,
-    maxWidth: 300,
-    wordWrap: 'break-word',
-  },
-
+const useStyles = makeStyles()({
   hoverVertical: {
     background: '#333',
     border: 'none',
@@ -38,7 +17,7 @@ const useStyles = makeStyles()(theme => ({
     position: 'absolute',
     pointerEvents: 'none',
   },
-}))
+})
 
 type Coord = [number, number]
 
@@ -65,37 +44,16 @@ const Tooltip = observer(function Tooltip({
   clientRect?: DOMRect
   TooltipContents: TooltipContentsComponent
 }) {
-  const theme = useTheme()
   const { featureUnderMouse } = model
   const { classes } = useStyles()
-  const { refs, floatingStyles, context } = useFloating({
-    placement: 'right',
-    strategy: 'fixed',
-  })
 
   const x = clientMouseCoord[0] + 5
   const y = useClientY ? clientMouseCoord[1] : clientRect?.top || 0
-  const clientPoint = useClientPoint(context, { x, y })
-  const { getFloatingProps } = useInteractions([clientPoint])
-
-  const popperTheme = theme.components?.MuiPopper
-
   return featureUnderMouse ? (
     <>
-      <Portal container={popperTheme?.defaultProps?.container}>
-        <div
-          className={classes.tooltip}
-          ref={refs.setFloating}
-          style={{
-            ...floatingStyles,
-            zIndex: 100000,
-            pointerEvents: 'none',
-          }}
-          {...getFloatingProps()}
-        >
-          <TooltipContents model={model} feature={featureUnderMouse} />
-        </div>
-      </Portal>
+      <BaseTooltip clientPoint={{ x, y }}>
+        <TooltipContents model={model} feature={featureUnderMouse} />
+      </BaseTooltip>
 
       <div
         className={classes.hoverVertical}
